Renamed from util to Flog. Cleaned out all of the junk modules.
