+++ src/test/scala/com/phasmidsoftware/flog/FlogSpec.scala
@@ -2,7 +2,7 @@
  * Copyright (c) 2019. Phasmid Software
  */
 
-package com.phasmidsoftware.util
+package com.phasmidsoftware.flog
 
 import org.scalatest.matchers.should
 import org.scalatest.{BeforeAndAfterEach, flatspec}
