+++ src/test/scala/com/phasmidsoftware/flog/LazyLoggerSpec.scala
@@ -1,4 +1,4 @@
-package com.phasmidsoftware.util
+package com.phasmidsoftware.flog
 
 import org.scalatest.matchers.should
 import org.scalatest.{BeforeAndAfterEach, flatspec}
