+++ src/test/scala/com/phasmidsoftware/flog/LoggablesSpec.scala
@@ -2,7 +2,7 @@
  * Copyright (c) 2019. Phasmid Software
  */
 
-package com.phasmidsoftware.util
+package com.phasmidsoftware.flog
 
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
